// src/ContentDashboard/Client/src/api/types.gen.ts
export type ContentOwnerResponseModel = {
    id: string;
    name: string;
    email?: null | string;
    documentCount: number;
};